pyproject.toml
[tool.poetry]
name = "poetry_fastapi"
version = "0.1.0"
description = ""
authors = ["zhouys <1767703660@qq.com>"]

[tool.poetry.dependencies]
python = "^3.8"
fastapi = "^0.78.0"
uvicorn = {extras = ["standard"], version = "^0.18.2"}
gino = {extras = ["starlette"], version = "^1.0.1"}
alembic = "^1.8.0"
psycopg2 = "^2.9.3"
gunicorn = "^20.1.0"
loguru = "^0.6.0"
APScheduler = "^3.9.1"
python-jose = "^3.3.0"
peewee = "^3.15.1"
PyMySQL = "^1.0.2"
casbin = "^1.16.9"
casbin-sqlalchemy-adapter = "^0.4.2"
nanoid = "^2.0.0"
redis = "^4.3.4"
passlib = "^1.7.4"
bcrypt = "^3.2.2"

[tool.poetry.dev-dependencies]
pytest = "^5.2"
requests = "^2.28.1"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
